backend/requirements.txt
flask
flask-cors
bcrypt
PyJWT
psycopg2-binary
langchain
langchain-google-genai
python-dotenv
openpyxl
invoke
gunicorn
sympy
